+++ packages/frourio/apis/users/_userId@number/index.ts
@@ -1,7 +1,0 @@
-import { UserInfo } from '../'
-
-export type Methods = {
-  get: {
-    resBody: UserInfo
-  }
-}