+++ packages/frourio/frourio.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  port: 10000,
-  basePath: '/api',
-  staticDir: 'packages/frourio/public',
-  cors: true,
-  dotenv: true,
-  immediate: false
-}